requirements/base.txt
asgiref==3.8.1
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.10
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.3.2  # https://github.com/redis/hiredis-py
python-decouple==3.8 
phonenumbers==8.13.49


# Django
# -----------------------------------------------------------------

Django==4.2.16
django-allauth==0.59.0  # https://github.com/pennersr/django-allauth
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-environ==0.11.2


# Django REST Framework
# -----------------------------------------------------------------

django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
dj-rest-auth==6.0.0
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
django-anymail==12.0
whitenoise==6.6.0

# installed for chat app
# -------------------------------------------------------------------------------
channels==4.1.0 
channels-redis==4.2.0 
daphne==4.1.2 